requirements.txt
pyserial
sqlite3
